+++ NEWS.md
@@ -1,5 +1,7 @@
 # R7 0.0.0.9000
 
+* `method_call()` is now `R7_dispatch()` (#200).
+
 ## Mar 2022
 
 * `convert()` allows you to convert an object into another class (#136).